requirements.txt
PyJWT==1.4.2
auth0-python==2.0.1
configparser==3.5.0
future==0.16.0
oauthlib==2.0.1
requests==2.13.0
requests-oauthlib==0.8.0
wsgiref==0.1.2
